feat: One Time Passwords
Added OTPs to verify newly signed-up user emails
and to offer the ability to reset forgotten passwords.

The reset of the password reuses the two password
input fields in the SignUpComponent that are now
refactored in their own component and might be
reused also when the ability to update the user
password is implemented.

Every mention of login, logout, and register has
been replaced with, respectively, sign-in, sign-out,
and sign-up.